chore: disable github workflows and actions
As part of incorporating the upstream changes, we want to disable their
workflows to prevent unintentional runs. Another way to do this is by
deleting the contents of those directories, but in this case we chose to
move them elsewhere because that reduces the possibility of conflicts if
these files were to be updated in upstream and then reintegrated here.