feat: add proof-of-sql-commitment-map trait with generalization over commitment schemes
As we develop the substrate "commitment pallet", we will need to store commitments or hashes of commitments on chain via the substrate key-value storage APIs.
The library that is introduced in this PR will provide an abstraction for commitment storage, especially the logic for table-wide operations that apply to all commitment schemes.

The commitment map abstraction ends up using some sum-types and product-types over commitment schemes as parameters.
It deals with enums that have variants of data per commitment scheme, and structs that have fields per commitment scheme.
It seems that the concept of algebraic data types over all commitment schemes will be repeating in this crate, and probably the commitment pallet in general.

For the commitment map in particular, different implementers will store different types (the test implementation will store TableCommitments, while the substrate implementation will use substrate-provided hashes).
I decided that it's for the best if these implementations declare what they store rather than hiding it, so that it can be left up to the caller to provide the correct data type.

So, in an effort to generalize the algebra of commitment schemes - a GenericOverCommitment trait is the first thing introduced here, which allows for generalizing over types that generalize over commitments.
Then, the commitment scheme trait provides the algebra - a sum-type AnyCommitmentScheme<T: GenericOverCommitment> and a product-type PerCommitmentScheme<T: GenericOverCommitment>.
Many types can be plugged into these, I implemented the trait for some common proof-of-sql types even though we won't use all of them here, since it's easy to do so.