feat: add CommitmentMap abstraction with test implementation
As we develop the substrate "commitment pallet", we will need to store commitments or hashes of commitments on chain so that they go through consensus, and are ultimately signed.
The substrate storage abstractions essentially provide a key-value store to accomplish storage operations like this.
In our case, the key will be (schema_name, table_name, commitment_scheme) and the value will be the commitment, or a hash of the commitment.

This PR provides an abstraction for representing mappings of this kind.
It implements many of the proof-of-sql-specific considerations.
In particular
- we want commitment scheme usage to be extensible to add more commitment schemes in the future
- we want to update all commitment schemes defined for a table simultaneously
- we don't want to update a table with commitment schemes that it was not originally defined with

This abstraction is tested by the test implementation MemoryCommitmentMap.