feat: add substrate implementation of CommitmentMap
As we develop the substrate "commitment pallet", we will need to store commitments or hashes of commitments on chain so that they go through consensus, and are ultimately signed.
The substrate storage abstractions essentially provide a key-value store to accomplish storage operations like this.
In our case, the key will be (schema_name, table_name, commitment_scheme) and the value will be the commitment, or a hash of the commitment.

A previousc commit added an abstraction for this in the form of a CommitmentMap, with additional proof-of-sql-specific considerations.
This commit introduces the implementation of this abstraction for substrate storage, in the form of the CommitmentStorageMapHandler.