feat: add on-chain-table crate
We need a type to represent insert data.
`RecordBatch`s would be ideal, but they cannot be understood in the
runtime due to a lack of `no_std` compatibility.
`proof-of-sql` `OwnedTable`s are an option that could be `no_std`
sooner, but they have internal-use-only column types and generics that
we do not need to deal with at a higher-level.
This crate will provide a middling type that can be converted to/from
`arrow` and be used by `proof-of-sql`.