feat: add OnChainColumn as no_std insert column type
The native `no_std` table type will essentially be a mapping of column
identifiers to column data.
`OnChainColumn` will represent that column data.
This opposes `arrow`'s `RecordBatch` which uses dynamic typing for its
column data.
An enum is used here instead, which is more similar to proof-of-sql's
`Column`, and offers a simpler implementation.