feat: add OnChainColumn::try_to_committable_column method
In order to use proof-of-sql's commitment API, you must use proof-of-sql
CommittableColumns, or another column type that implements conversions
to CommittableColumns.
This method provides a conversion is generic over a scalar, so it cannot
use the typical conversion traits unless a new type is created.
It's also fallible in the case that a decimal is out of the scalar's
bounds, which can happen as the construction of this column does not
validate decimal values.