build: use rust 1.81.0 and fix new clippy lints
In rust 1.81.0, the Error trait was stabilized in core.
This upgrade is important to use some dependencies with new no_std
support that depend on the Error trait.

This introduces new clippy lints that fail within pallet macros.
To address this, pallet modules are now configured to allow these lints.