feat: promote optional PerCommitmentScheme fields to higher kinded types
Currently, PerCommitmentScheme's fields are all wrapped in Options.
In the commitment-map code, it was only used in where it made sense to
type a collection of at most one data item per commitment scheme.
However, in order for it to be a true product type over commitment
schemes, this Option wrapping should be removed.
Instead, PerCommitmentScheme<T>s should be re-typed with the new
OptionType as PerCommitmentScheme<OptionType<T>>.