feat: add commitment scheme mapping API
Similar to rust's Option::map, it is useful to be able to map a
AnyCommitmentScheme<T> to an AnyCommitmentScheme<U>.
The same goes for PerCommitmentScheme, but that is more analagous to
Iterator::map.
This introduces AnyCommitmentScheme::map and PerCommitmentScheme::map.

In concept, these map APIs accept a higher-order function.
This function is generic over C, accepts a T::WithCommitment<C>, and
returns a U::WithCommitment<C>.
The result of AnyCommitmentScheme::map keeps the original value, just
mapping its internal value generically.
The result of PerCommitmentScheme::map maps every field respectively.

However, higher-order generic functions are not supported in rust, so
instead a trait GenericOverCommitmentFn has been introduced.
Types that implement this trait will implement its generic call
method, and can then be supplied to the map APIs.