feat: allow non-generics in commitment scheme types
It can be useful to define AnyCommitmentSchemes and PerCommitmentSchemes
where the value isn't actually generic over the commitment scheme.
Instead, the value is just logically associated with the commitment
scheme, not bound to it by type.
This is especially useful when using the map APIs to produce a
non-generic value, e.g., getting the range of a TableCommitment.

To this end, a ConcreteType has been added.
This implements GenericOverCommitment, but it's the degenerate case.
An unwrap method leveraging this has been added to AnyCommitmentScheme.