feat: add column type conversion utilities
To be able to begin computing commitments for a table, its schema needs
to be understood in proof-of-sql types.
Since the table definitions will use sqlparser types, this necessitates
a sqlparser data type to proof-of-sql column type conversion.
This conversion will also act as a validation layer, since not all
sqlparser types are supported by proof-of-sql.