feat: define row number metadata column
For proof-of-sql, it is important for the prover to operate on the rows
of data in the same order that they were committed to.
In order to satisfy this requirement, a row number metadata column will
be added to tables during commitment computation.
This column's order will be equal to the order that the rows were
committed to, so provers merely need to ORDER BY this column.