feat: validate table definitions for proof-of-sql
Sqlparser's table definition structure is much more permissive than what
proof-of-sql can support.
If a DDL is accepted, we should be able to commit to and perform
proof-of-sql queries on the resulting table.
These changes introduce utilities to validate a sqlparser
CreateTableBuilder, so that we can catch these issues early.
It does so in the form of a wrapper type ValidatedCreateTable, which may
provide some nice type guarantees to functions that depend on these
requirements.