feat: store unhashed table commitment bytes
Currently, the CommitmentStorageMap types intend to store a hash of a
TableCommitment as their value.
However, storing a hash of the TableCommitment instead of the raw
TableCommitment introduces a lot of complexity.
In particular
- the commitment data model in the sql database would need to include
  commitments, and any other data necessary to update commitments
- the commitment pallet would depend on data in the sql database to
  update commitments
- verification workflows would also need to connect to both substrate
  and the sql database to obtain verifiable commitments

To reduce complexity, this changes the CommitmentStorageMap to raw table
commitments.
This comes at the cost of storage size, but the simplicity is worth the
overhead.
The correct path to optimize the size probably isn't hashing anyway -
it's "single table commitments" in proof-of-sql.