refactor: split commitments from commitment-sql return structs
APIs will be added to the commitments pallet that correspond to the
top-level functions of the commitment-sql crate.
These APIs will call these functions, consume their returned commitments
into storage, then return the remainder of the function's return values.
Without defining a new struct, the easiest way to facilitate the partial
consumption of these return types is to split them into a tuple.