feat: process DDL and store snapshot commitments in pallet-commitments
pallet-commitments currently has support to process regular DDL, but
another DDL use case includes initial data snapshots.
These data snapshots will come with pre-computed commitments, so
pallet-commitments shouldn't initially store empty ones.
However, the rest of the DDL needs to undergo normal processing and
validation.
The logic for this processing was previously added statelessly in the
commitment-sql pallet.
This introduces an API on pallet-commitments exposing this processing
and initiating the snapshot commitments in storage as a side effect.