feat: configure dory sigma with genesis config
A value called "sigma" is used when creating a dory public setup.
This value has some nuances to it regarding performance.
If it's smaller, the entire public parameters can be smaller, so making
it small leaves room to make the setup generation very quick.
However, it also affects the max table size that can be committed to,
with larger values facilitating larger tables.

All of that is to say that it's best for sigma to be configurable.
Unit tests should prefer to use small values while production should go
higher.