test: lower max_nu and sigma for dory setup in unit tests
The current values aren't very large, but they are large enough to slow
down unit tests significantly.
The tables used in unit tests are all quite small, so we can affort to
set the default values for max_nu and sigma much lower.
With this change, pallet-commitments tests take 2 seconds to run instead
of 18.