feat: statically use proof-of-sql setups from file instead of storage
Originally, we had the proof-of-sql public parameters placed in storage,
and were going to initialize them with genesis config.
However, I'm anticipating this storage retrieval to be a bottleneck.

We also need to load the public-parameters from a file in some way, to
avoid having to regenerate it and to share with provers/verifiers.

This change alleviates the storage bottleneck by just baking the public
parameters into the runtime binary w/ lazy_static.
It also accomplishes "loading" from a file with include_bytes.

During testing, I saw a 12x improvement in test speed when compared to
testing with the same large public parameters placed in storage.