feat: clamp decimal precision to proof-of-sql maximum
Proof-of-sql only supports decimal precision up to 75.
The ethereum core ddl, on the other hand, contains decimal precision up
to 78.
These columns generally do not store values this high, but in order for
proof-of-sql to commit to them we need to claim that they fit in
precision 75, in both ddl and dml.
This introduces utilities for making this claim.