fix: preserve record batch timezone existence in on-chain-table
While columns with/without timezones are now both allowed in DDL, the
record-batch and on-chain-table round-trip doesn't preserve the timezone
correctly.
In particular, if the timezone didn't exist as inserts enter the chain,
it will exist as they are inserted to flightsql.
Then, flightsql will use TimestampTz for these, instead of Timestamps.

Making the timezone optional in on-chain-table, and only forcing it to
UTC for commitment, allows the timezone type info to be preserved
throughout the lifetime of an insert.