feat: apply data compliance with target types to realtime data
We do force realtime data to compliance with proof-of-sql and postgres
today, as well as historical data.
However, some of these transformations require knowledge of target
types, the source of truth for that being the table schema.
Applying these transformations to realtime data requires accessing the
table's schema in the native APIs.

This changes the native apis to accept the table definitions stored on
chain, and also perform any necessary transformations.
In particular, if a column is supposed to be a decimal but it's a
string, we perform the parsing now on realtime data.