fix: coerce insert data column order into existing commitment order
Currently, the runtime shares the proof-of-sql requirement that data
appended to commitments needs to have the same column order as the
commitments.
However, this is a UX issue, and is currently preventing some garfield
inserts in testnet.
This PR loosens that requirement by coercing insert data into the same
order as the existing commitments before committing and QuorumReached.