feat: benchmark indexing pallet
Currently, the heaviest pallet in terms of compute and memory is the
indexing pallet, but all of its weights are hardcoded at zero.
This makes it difficult for the chain to determine how to divide up its
transactions between blocks, and we think this may cause blocks that are
too expensive.
The benchmarking added here will help generate weights for these
transactions, which should help alleviate this issue.