fix: adjust NativeApi trait bounds and fix docs in pallet-indexing
The pallet-indexing config has a trait bound requiring that I implements
NativeApi.
This trait bound leads to some bugs in the #[frame_support::pallet]
macro, breaking doc builds.
The current workaround for this was disabling various pallet-indexing
modules in doc and doctest builds.

It seems, however, a simpler workaround is to just not have that trait
bound on the pallet-indexing config at all.
Instead the trait bound can just be applied wherever it is used, and it
already is.