feat: wrap TableCommitmentBytesPerCommitmentScheme in new PassBy type
As more commitment computation code is moved to the native interface, it
will become important to be able to have commitment data cross that
interface boundary. TableCommitmentBytesPerCommitmentSchemePassBy will
be the vehicle to accomplish this.

It was necessary to use a wrapper type in this situation as the
PassByCodec derive macro does not like generic types.
TableCommitmentBytesPerCommitmentScheme is merely a type alias for a
deeply generic PerCommitmentScheme type.