feat: define machinery supporting table-defined privileged quorum
We want to be able to provide certain indexers with special access to
certain tables. The proposed design introduces an additional "quorum
scope", that can only be contributed to by indexers that have a
new table-specific permission. This change introduces new types to
sxt-core that will be involved in this design:
- QuorumScope, for identifying public vs privileged quorum
- InsertQuorumSize, for table-defined quorum sizes per quorum scope
- SubmittersByScope, for tracking agreeing submissions per quorum scope
- a new SubmitDataForPrivilegedQuorum permission that is table-specific