test: use common table definition in pallet-indexing tests
The pallet-indexing tests have some repeated code between eachother
redefining the table being indexed to, as well as some code that is
redundant to pallet-tables API for creating the table. This change
introduces a common utility function for defining a sample table, and
makes all tests create the table using the appropriate pallet-tables
function.