feat: upgrade proof-of-sql to 0.73.1
proof-of-sql 0.73.0 is not no_std compatible. There was a bug in the CI
that allowed no_std incompatibilities to build up. Most of these have
been addressed in 0.73.1.

The one no_std incompatibility remaining is proof-of-sql's use of the
serde feature of sqlparser. This feature is not no_std compatible,
though there's nothing preventing that compatability except for Cargo
misconfiguration. So, this change also uses a sqlparser patch with Cargo
configuration changes for the time being.

This latest version of proof-of-sql also includes the VarBinary type,
which is critical for our next version of indexed tables.

Co-authored-by: Trevor Lovell <tlovell-sxt@spaceandtime.io>