feat: add off-the-shelf pallet-multisig to runtime
We want to have a multisignature/voting mechanism for performing
system-level operations in a more decentralized way than plain sudo.
This requirement is actually fulfilled by a composition of
pallet-multisig and pallet-sudo.

The multisig pallet provides higher-order transactions and handles
voting state for them. When the defined threshold of signatures has been
met, it performs a transaction as a multisignature account, whose key is
determined as a function of the signatories and threshold for the
transaction. We can know what the multi-sig account key would be for the
sudo "parliament" we want to "elect", so we can just set the sudo key to
that multi-sig key and then we're done.

Notably, we don't have to do this at genesis, we can have a single sudo
account for some time and, when we're ready, that account would be
responsible for handing the key off to a multisig account.