fix: 0-index META_ROW_NUMBER column
The META_ROW_NUMBER column was originally designed to be 0-indexed.
However, we changed it to 1-indexed after historical parquet data for
the column was generated as 1-indexed. Now that we are recreating
historical data, we can make it 0-indexed again, which will be an
important property now that the prover assumes that META_ROW_NUMBER has
no gaps.