feat: feature-flag multiple static-setups nu sizes
After switching to native commitment generation, the only commitment
operations being performed in the runtime are commitments to empty
tables (to initialize commitments for new tables). This means that the
runtime doesn't need a very large setup to perform all of its commitment
computation. This change makes it so a tiny setup with nu 1 is used in
the runtime while the current setup with nu 14 is used in native.