feat: increase large proof-of-sql static-setup to nu 15
The nu value of the setup used to calculate proof-of-sql commitments
determines the maximum number of rows in a table. Increasing from nu 14
to nu 15 increases the maximum number of rows from 67 million to 268
million. This can be done without increasing the runtime binary size
because the runtime only uses nu of 1.