feat: alias ByteA for postgres compatibility
Currently, all aliases for bytes columns are supported except one:
postgres's bytea. In the future we will remove all aliases that are not
supported in the space and time data warehouse backend. When that time
comes, this alias will be removed again, but while postgres is still the
backend, we need to support its column types.