feat: define proof-of-sql setup loading configuration and functionality
Proof-of-sql setups are currently baked into the native and runtime
binaries. This is not strictly necessary for the native code, as native
code has access to io. This will be especially important as we add
hyperkzg support, since hyperkzg setups can be quite large.

This defines a clap struct for proof-of-sql setup loading configuration.
It also adds a function for performing this loading for dory.