feat: store loaded proof-of-sql setups statically
Proof-of-sql setups are currently baked into the native and runtime
binaries. This is not strictly necessary for the native code, as native
code has access to io. This will be especially important as we add
hyperkzg support, since hyperkzg setups can be quite large.

This defines static OnceCells that loaded proof-of-sql setups will be
stored in. Native interfaces cannot access values created outside of the
interface unless they are simply static. This also adds functions for
initializing these cells from configuration.