feat: use io static setups in native and initialize them on startup
Native commitment code currently uses proof-of-sql setups that are
currently baked into the binary. This is not strictly necessary, as
native code has access to io, and can use the newer loaded static
setups.

This change updates the native code to use the loaded static setups and
initializes them on startup accordingly.