refactor: remove nu_15 and give remaining "baked" setup its own module
Now that native code uses the io proof-of-sql static setups, the nu_15
feature of that crate is unused. This change removes the nu_15 feature,
and renames the remaining nu_1 static setup to "baked". It also gives
the baked static setup its own module to match the structure of the io
alternative.