fix: guarantee column identifiers are uppercase in OnChainTable
Previous versions of proof-of-sql had case-insensitive identifiers.
However, as proof-of-sql switches to using sqlparser/datafusion types,
proof-of-sql is being designed with more case sensitivity. The
responsibility for using proof-of-sql in a case-insensitive way is being
pushed downstream.

This change addresses this issue in a simple way. Make OnChainTable
case-insensitive as a type guarantee. All constructors of OnChainTable
will now coerce to uppercase. This will make all creation of
proof-of-sql commitments use uppercase as well, since all of sxt-node
computes commitments from OnChainTable data.