feat: upgrade to proof-of-sql 0.80.2
Proof-of-sql has made some changes around the hyperkzg commitment
scheme that are required for us to be able to use it in sxt-node. In
particular, hyperkzg commitments (not proofs) are now no_std compatible,
type aliases have been provided for its setups, as well as
deserialization utilities. This change upgrades proof-of-sql so that we
may use hyperkzg soon.