test: add TestDirectory test utility to static-setups crate
The static-setups crate will soon be home to some functionality that has
side effects on the file system. In order to test this functionality, it
will help to have some method of quarantining side effects on a per-test
basis.

This change adds such a utility. On construction, a directory for the
current test is randomly generated. On drop (i.e. test pass or test
failure), the directory is deleted.