feat: download hyperkzg public setup files
We want to be able to download hyperkzg public setup files on startup.
This adds a function for doing so that accepts a destination directory
for the public setup files to be downloaded to, and a "release degree"
parameter that corresponds to a proof-of-sql github release of
compressed ppot files. If the files that would be downloaded from the
github release already exist, they are skipped.