feat!: replace ipa with hyperkzg
The HyperKZG commitment scheme is a relatively recent addition to
proof-of-sql. It is most important for on-chain verification allowing
for the verification of proof-of-sql proofs in the evm. We'd like to
support such a use case, so we need to add it as a commitment scheme in
sxt-node.

This doesn't just add it though, it replaces ipa. The IPA commitment
scheme has thus far been unused in sxt-node, and doesn't have any
benefits over dynamic dory. The main reason it was included in the
first place was to demonstrate the correctness of the commitment scheme
abstractions.