fix: keep file bytes independent during hyperkzg load and use rayon
Some situations seem to cause the load of larger multi-file hyperkzg
setups to hang. The load functions do some reorganization of the bytes
for the sake of convenience and parallelization. First, the buffers
holding the individual file bytes are flattened into a single buffer,
and then split up again into multiple buffers according to the cpu count
to be deserialized in parallel.

The hang bug was resolved locally by instead keeping the bytes in their
original file-separate buffers and deserializing them one file at a
time. This left a lot of cpu resources on the table, so the
deserialization was re-parallelized again using the appropriate
cpu-bound library rayon. However, now parallelization is only performed
within files, rather than between files. This is much simpler, and also
reduces the memory usage during deserialization significantly, as
explained in the comments.