feat: add attestation rpcs
We'd like to verify that the network has come to agreement on certain
slices of storage, and we'd like to be able to do so across languages,
especially solidity. That second requirement compelled us to create the
attestator architecture. However, in order to support easily grabbing
attestation signatures out of chain storage across languages, we also
need to provide read-only RPCs.

This change introduces two new RPCs, one for getting the attestations
for any given block, and the other for conveniently getting the "best
recent attestations".