fix: enable native-api/std by pallet-commitments/std
When compiling pallet-commitments with the "std" feature, and without
any other workspace-level context, compilation errors occur. This is
because pallet-commitments/std depends on some items gated behind
native-api/std. This is easy to miss at the workspace level because the
resolver ends up enabling native-api/std due to some other crate
requiring it.