feat: add attestation_tree crate
The current watcher/event-forwarder crates use bespoke merkle trees to
multi-sign small slices of chain state. This is important for verifying
account balances and commitments in the EVM. However, as part of the
on-chain query path, another component will be using the same merkle
trees - an RPC.

Naturally, we would like to reduce as much repeated code as possible
between this new RPC, the watcher crate, and the event forwarder crate.
One way in which the current merkle tree utilities do not suffice is
that they tend to rely on the use of subxt clients. We would like to
move away from using these clients in this repository, and the RPC will
not be using them. To address this, this crate will provide as much
common logic as it possibly can in a stateless manner.

We have also wanted to design the merkle tree utilities to provide more
abstraction. It would be nice to be able to arbitrarily tell the
utilities what storage prefixes to include, and how to encode them into
leaves. This crate will provide an initial version of these
abstractions.