feat: add HashAndKey and HashAndKeyTuple traits to attestation_tree
The attestation_tree crate will provide abstractions for defining how
keys/values from a given storage prefix should be encoded into leaves.
It would be nice if these leaf-encoding definitions had access to the
strong types of the prefix. So, they should have the types for the keys
and values of the storage prefix as associated types.

However, the storage prefix and keys alone are not enough to be able to
decode a raw storage key. You also need to know what hashing algorithms
were used on the storage keys.

This adds the HashAndKey trait, which is implemented on pairs of
(hasher, key), allowing us to express an association between a key and
its hashing algorithm in a storage instance.

This also adds a HashAndKeyTuple trait, implemented on tuples of up to 4
HashAndKeys, as storage instances can be defined with a variable number
of keys, or none at all. This trait has a bit more functionality,
providing an associated type of just the key tuple, as well as encoding
and decoding functions.