feat: add CommitmentMapPrefixFoliate to attestation_tree
This adds a new PrefixFoliate implementor for encoding the on-chain
commitment storage into the attestation tree. It customizes the encoding
of these types heavily for ease of use in solidity, as well as all the
other pieces of the on-chain verification architecture. Most notably,
the TableIdentifier is stringified, as table-identifiers-as-strings is
the standard among most sxt APIs.