feat: add merkle tree creation and proving utilities to attestation_tree
The attestation tree can now be created, and leaves can be proved,
leveraging the attestation_tree abstractions. Ideally, we'd like to make
it so adding more prefixes to the list would be implicit downstream, but
this would require much more dynamic typing, unfortunately. As it is,
callers need to know what prefixes to grab from state, but everything
else is performed by these attestation_tree utilities.