feat: add proof plan RPCs
A barrier to usage of proof-of-sql against chain data is the creation of
proof-of-sql proof plans. Proof plans are similar to normal sql logical
plans, but they are specifically for proof-of-sql proving/verification.
The input for creating a proof plan, at a high level, is query text and
schema information for all tables in a query.

The chain has proof-of-sql schema information for tables inside queries,
so it makes sense to provide proof plan APIs on the chain itself. This
change adds a couple RPC calls for this - proofPlan and evmProofPlan.
The former supports more queries while the latter can be verified in the
evm.

As much of this logic as possible has been implemented in a stateless
way, and all state is accessed via runtime APIs.