feat: add pallet_migrations to runtime
We need to incoroporate the polkadot migrations pallet into our runtime
if we want to be able to perform multi-block storage migrations. This is
especially relevant as we are entering a more stable development phase,
and all breaking storage changes in the future will need to be handled
via migrations.