fix: use environment variable for validator name
To make it more evident to users that this value needs to be set, an
environment variable will be used, and this will break if it has not
been set. Since this is not really an option in docker-compose it has
been removed there. The node will still generate a name upon startup,
and it's shown in the logs.