chore: remove flightsql-pg from docker image
FlightSQL is no longer needed, however, the main reason to remove this
is to get rid of the ARTIFACTORY_LOGIN secret because we'll change how
we log into artifactory soon. Other related environment variables have
been removed as well, the code that reads them has a default value for
them, except DATABASE_URL, so it's kept there to reduce the chance of
things breaking.