feat: add runtime api for reading commitments storage for any scheme
When performing a query, it can be important to get all the commitments
for all the tables in a query, and to reject if any tables are missing
commitments. There are even some use cases where you want to get all the
commitments for a query and the commitment scheme doesn't matter, like
building a proof plan.

This adds a runtime API around the new pallet-commitments utility for
reading commitment storage for multiple tables and any scheme. This will
eventually be used by RPCs to produce proof plans.