feat: read commitments storage for multiple tables and any scheme
When performing a query, it can be important to get all the commitments
for all the tables in a query, and to reject if any tables are missing
commitments. There are even some use cases where you want to get all the
commitments for a query and the commitment scheme doesn't matter, like
building a proof plan. This adds methods to the commitments pallet for
performing these kinds of reads, which will eventually be used by
runtime APIs/RPCs to produce proof plans.