feat: add CommitmentScheme::into_any_concrete method
There are utilities for converting from `AnyCommitmentScheme` types to
its unit-variant alternative `CommitmentScheme`, but not the other way
around. The other way around is a little stranger since
`AnyCommitmentScheme` can store information, whereas the current
utilities are just lossy. This solution is similar to the standard
library's `bool` -> `Option` conversion `then_some`, where a value for
the variant is provided as input. This can be useful when writing code
that wants to iterate through commitment schemes using
`GenericOverCommitmentFn` utilities.