fix: limit memory allocation when deserializing proof plans
By default, bincode does not limit memory allocation during decoding. If
a large number is misinterpreted as a length specifier for a collection,
bincode will attempt to allocate a massive collection accordingly. We
observed this by chance when using a normal DynProofPlan in the
verifiableCommitmentsForProofPlan RPC. It attempts to deserialize the
proof plan as an EVMProofPlan first, and in this DynProofPlan happened
to run into this problem when interpreted as an EVMProofPlan (not all of
them do).

The solution is to configure bincode with limits. We already limit the
proof plan's byte size during input, and if it's valid, its in-memory
representation shouldn't be larger than its serialized representation
since we use a pretty not-compact bincode config. Now, we get errors
instead of panics.