feat: impose bound on number of batches finding quorum
There is no mechanism as we speak to expire batches that were never able
to find quorum. This will eventually lead to storage becoming
unmanageably large, and bad actors (if they had permission to) could
attack this unbounded storage with submissions that will intentionally
never reach quorum. A basic protection against this would be to prune
batches that are still finding quorum after a certain limit has been
reached. This change accomplishes this, keeping track of batch
submission order with a new BatchQueue and attempting to prune
submissions from the back of the queue once a configurable capacity is
reached.