feat: migrate storage to populate batch queue from submissions
Recently, the indexing pallet imposed a bound on batches finding quorum
with the help of a new BatchQueue in storage. This adds a storage
migration to populate the initial contents of this BatchQueue, by
analyzing the submissions storage.

This implementation is simple but dumb. There's no way for us to know
what order the batches were initially submitted in from current storage,
so this just inserts them to the batch queue in their iteration order
(lexicographic ordering of the hash of the batch ids). Furthermore, this
does not do any pruning, so ideally this the batch queue capacity should
exceed the number of batches finding quorum in all live networks so that
they are pruned more gradually.