feat: add proof-of-sql-unversioned crate
We will need the ability to deal with multiple versions of proof-of-sql
in sxt-node in the future. We will need to have old native code use
older versions of proof-of-sql forevermore, or just come up with some
solution with newer versions that matches legacy behavior. We will
sometimes need to upgrade the proof plan functionality of RPCs more
often than we upgrade the commitment functionality.

Some sxt-node code will be philosophically version agnostic, so is
sort-of shareable between downstreams w/ different proof-of-sql
versions, but not as regular rust code since that would require defining
a particular version. This new crate provides a place to put code to
share between downstreams with disparate proof-of-sql versions, and also
provides the first module of this variety: converting sqlparser types to
proof-of-sql types. This code will now be shareable between the RPCs and
native code, and even indexers, who may use different proof-of-sql
versions.