feat: convert create statements to communicable table schemas
Some uses of the tables pallet are only interested in table's schemas,
as in, their column names and data types. However, we store table
schemas as SQL create statements. We would like to provide a runtime API
for reading this storage into a more structured schema type. This change
provides some types and functions to sxt-core that will be helpful in
this effort.