feat: add table_schema method to pallet-tables
Some uses of the tables pallet are only interested in table's schemas,
as in, their column names and data types. However, we store table
schemas as SQL create statements. We would like to provide a runtime API
for reading this storage into a more structured schema type. This change
provides a stateful method for performing this read operation and
processing. It can be used to implement the future runtime API easily.