refactor: use unversioned column type conversion in commitment-sql
The code for defining what column types we support and how they convert
to proof-of-sql types has recently been moved to the
proof-of-sql-unversioned crate. This will allow it to be shared with
different downstreams that may use different versions of proof-of-sql.
This change completes the repeated code reduction by using the
unversioned macros in commitment-sql.