feat: implement column type conversion in rpc crate
The proof plan RPCs require knowledge of table schemas in storage.
Currently, they get this from the commitment metadata for tables in
storage, but we want them to get these schemas from the tables pallet
metadata instead. In this path, we get the data types as plain sql,
which we need to parse and then convert to proof-of-sql column types.

This implements the necessary column type conversion code in the rpc
crate. So that we will be able to upgrade the proof-of-sql version of
RPCs separately from the runtime, we use the proof-of-sql-unversioned
macro to generate a new implementation instead of using an existing one.