feat: add QuerySchema schema accessor
The proof plan RPCs require knowledge of table schemas in storage.
Currently, they get this from the commitment metadata for tables in
storage, but we want them to get these schemas from the tables pallet
metadata instead.

proof-of-sql relies on the SchemaAccessor trait for getting schemas,
which is already implemented for QueryCommitments - mappings of table
refs to commitments. Now, instead of commitments, we'll have table
schemas from the tables pallet APIs. This adds QuerySchema, which can
be built from these table schemas, and implements SchemaAccessor.