refactor: use schemas from tables pallet in proof plan RPCs
The proof plan RPCs require knowledge of table schemas in storage.
Currently, they get this from the commitment metadata for tables in
storage, but we want them to get these schemas from the tables pallet
metadata instead. This change performs the switch in the RPC entrypoint,
and removes some now-dead code.